Moved irrelevant tests and functions from calib module to 3d.
